Added functions to js.script and created GitHub repository
An add function: this should add two numbers (parameters) and return the sum.
A subtract function: this should subtract the second number from the first number and return the difference.
A multiply function: this should multiply two numbers and return the product.
A divide function: this should divide the first number by the second number and return the quotient; however, if the divisor is equal to zero, the function should return the text “Not Allowed.”